server and client part: sort by reviews

+++ server/models/models.js
@@ -18,7 +18,7 @@ const BasketDevice = sequelize.define('basket_device', {
 const Device = sequelize.define('device', {
     id:{type:DataTypes.INTEGER, primaryKey: true, autoIncrement: true},
     name:{type:DataTypes.STRING, unique: true, allowNull: false},
-    price:{type:DataTypes.INTEGER, allowNull: false},
+    price:{type:DataTypes.FLOAT, allowNull: false},
     rating:{type:DataTypes.FLOAT, defaultValue: null},
     img:{type:DataTypes.STRING, allowNull: false}
 })
